temp
Former-commit-id: 42c0f59c3f2ffc268ec106f7cf8d64bb138da3c6

+++ common/datablock.py
@@ -71,7 +71,7 @@ class Datablock(object):
     def add_images_for_cluster(self, images, cluster_topic, partition=False, partition_index=0, num_partitions=1):
         images_in_cluster = get_images_for_cluster(images, cluster_topic)
         if(partition):
-            images_per_partition = len(images_in_cluster) / num_partitions
+            images_per_partition = int(len(images_in_cluster) / num_partitions)
             start = partition_index * images_per_partition
             end = partition_index * images_per_partition + images_per_partition
             images_in_cluster = images_in_cluster[start:end]
